+++ core/headers/Image.h
@@ -13,26 +13,26 @@
 class Image {
 public:
 	Image();
-	Image(float w, float h);
-	Image(float w, float h, float* c);
-	Image(float w, float h, float r, float g, float b);
+	Image(int w, int h);
+	Image(int w, int h, float* c);
+	Image(int w, int h, float r, float g, float b);
 	virtual ~Image();
 
 	void draw();
 	void draw(Position* p);
 
-	float getHeight() const;
-	void setHeight(float height);
-	float getWidth() const;
-	void setWidth(float width);
+	int getHeight() const;
+	void setHeight(int height);
+	int getWidth() const;
+	void setWidth(int width);
 	const float* getColor() const;
 	const float getColor(int i) const;
 	void setColor(float r, float g, float b);
 
 
 protected:
-	float width;
-	float height;
+	int width;
+	int height;
 
 	float color[3];
